# api/openapi.yaml
openapi: 3.0.3
info:
  title: corps (团队) API
  version: 0.6.0
  description: >
    MVP API for the corps multi-tenant collaboration SaaS. This document is the
    single contract for frontend and backend in Phase 3. All paths are prefixed with
    /api/v1/. Every response uses the uniform envelope { code, data, message }.
    Multi-tenant access control: every {wid} route verifies the caller's membership
    row for that workspace, then runs all tenant queries inside a transaction with
    app.workspace_id / app.user_id set (RLS binds at engine level in hardened
    deployments — see db/schema.sql).
    This contract currently defines 57 paths (updated 2026-09-09).
  contact:
    name: MVP Dev Expert Team — Architecture
servers:
  - url: http://localhost:3000
    description: Local development (Next.js Route Handlers)
  - url: https://ghcr.io/{env}/corps
    description: GHCR Docker deployment
    variables:
      env:
        default: staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{wid}/analytics/overview:
    get:
      tags: [analytics]
      summary: Workspace analytics overview (owner/admin)
      description: >
        Owner/Admin only (member → 403). Aggregates the last 14 days of workspace
        events (RLS-scoped): funnel step user counts with conversion rates,
        per-day event counts, top-8 event names, total event count and distinct
        active users.
      operationId: getAnalyticsOverview
      parameters:
        - $ref: '#/components/parameters/Wid'
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  data:
                    type: object
                    properties:
                      range:
                        type: object
                        properties:
                          days: { type: integer, example: 14 }
                          since: { type: string, format: date-time }
                      funnel:
                        type: object
                        description: Two-stage serialized funnel (acquisition + activation, D2 fix)
                        properties:
                          acquisition:
                            type: array
                            description: Acquisition funnel steps (grouped by sessionId)
                            items:
                              type: object
                              properties:
                                name: { type: string }
                                label: { type: string }
                                users: { type: integer }
                                rate: { type: integer, description: Conversion vs previous step (%) }
                          activation:
                            type: array
                            description: Activation funnel steps (grouped by userId)
                            items:
                              type: object
                              properties:
                                name: { type: string }
                                label: { type: string }
                                users: { type: integer }
                                rate: { type: integer, description: Conversion vs previous step (%) }
                      daily:
                        type: array
                        items:
                          type: object
                          description: '{ date, _total, <eventName>: count }'
                      topEvents:
                        type: array
                        items:
                          type: object
                          properties:
                            name: { type: string }
                            count: { type: integer }
                      totalEvents: { type: integer }
                      eventsTruncated: { type: boolean, description: Whether event sampling was truncated due to limit }
                      eventSampleSize: { type: integer, description: Number of events actually loaded (may be less than totalEvents) }
                      activeUsers: { type: integer }
                      coreActiveUsers: { type: integer, description: Distinct users with CORE_EVENTS in the window }
                      waw:
                        type: object
                        description: Weekly active users (current workweek, work-scoped)
                        properties:
                          weekStart: { type: string }
                          users: { type: integer }
                      retention:
                        type: object
                        description: D1/D7/D30 retention rates
                        properties:
                          d1:
                            type: object
                            properties:
                              active: { type: integer }
                              eligible: { type: integer }
                              rate: { type: integer, description: Retention rate (%) }
                          d7:
                            type: object
                            properties:
                              active: { type: integer }
                              eligible: { type: integer }
                              rate: { type: integer, description: Retention rate (%) }
                          d30:
                            type: object
                            properties:
                              active: { type: integer }
                              eligible: { type: integer }
                              rate: { type: integer, description: Retention rate (%) }
                      sessions: { type: integer, description: Distinct sessionId count in the window }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/workspaces/{wid}/billing/checkout:
    post:
      tags: [billing]
      summary: Create a Stripe Checkout Session (upgrade)
      description: >
        Owner only (AC-05). Creates a Stripe Checkout Session (mode=subscription) for the
        workspace plan; quantity is derived from `Workspace.seatLimit`. Returns the redirect
        URL. `priceId` defaults to env STRIPE_PRICE_ID. On checkout.session.completed the
        webhook writes the Subscription row and sets plan='pro'.
      operationId: createCheckout
      parameters:
        - $ref: '#/components/parameters/Wid'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session created (redirect URL)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Checkout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '501':
          $ref: '#/components/responses/PaymentNotConfigured'
...
components:
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT (15 min) in Authorization header. workspace_id resolved from JWT claim `wid`.
    refreshAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: >
        Better Auth session cookie (httpOnly, secure, 7-day). The /api/v1/auth/refresh
        endpoint reads this cookie to re-issue a workspace-scoped access JWT.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: HMAC-SHA256 signature of the raw body using the Stripe webhook secret.
  parameters:
    Wid:
      name: wid
      in: path
      required: true
      description: Workspace id. Must equal the JWT `wid` claim; mismatch is 403/404.
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 400
            data: null
            message: "email must be a valid address"
    Unauthorized:
      description: Missing or invalid token (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 401
            data: null
            message: "invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted (RBAC / tenant mismatch) (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 403
            data: null
            message: "role 'member' cannot perform this action"
    NotFound:
      description: Resource unknown to caller (RLS-filtered) (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 404
            data: null
            message: "workspace not found"
...
    PaymentNotConfigured:
      description: Payment provider not configured (501)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 501
            data: null
            message: "payment provider not configured"
    InternalError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 500
            data: null
            message: "internal server error"
...
  schemas:
...
    Envelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: "Business status code, mirrors HTTP status code (200=success, 4xx/5xx=error)"
          example: 200
        data:
          nullable: true
        message:
          type: string
          description: "Human-readable status message. Required for error responses (4xx/5xx), optional for success responses (2xx)."
          example: "ok"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 401
        data:
          nullable: true
          example: null
        message:
          type: string
          example: "invalid or expired access token"
        # L4：events POST 400 等校验失败响应会附带 zod errors 数组，
        # 此前未在 ErrorEnvelope 中文档化。这里补为可选字段（未列入 required），
        # 既覆盖现有带 errors 的 4xx 响应，也不破坏只返回 {code,message,data} 的错误响应。
        errors:
          type: array
          description: >
            Optional field-level validation errors (e.g. zod issues) for 4xx
            responses. Absent on non-validation errors.
          items:
            type: object
            description: A single validation issue (path/message/code as produced by zod).
...
    Envelope_Checkout:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/CheckoutSession'
...
    CheckoutRequest:
      type: object
      properties:
        priceId:
          type: string
          description: Stripe Price ID. Defaults to env STRIPE_PRICE_ID when omitted.
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
        period:
          type: string
          enum: [monthly, yearly]
          description: Billing cycle. Defaults to monthly when omitted; invalid values are rejected with 400.
        provider:
          type: string
          enum: [stripe, wechatpay-native, alipay-page]
          description: Payment channel. Defaults to env PAYMENT_PROVIDER when omitted; explicit value overrides the default for frontend payment-method selection.
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Redirect URL for Stripe Checkout / Alipay gateway / Billing Portal. Present for stripe and alipay-page channels; absent for wechatpay-native (use qrCodeUrl instead).
        qrCodeUrl:
          type: string
          description: WeChat Pay code_url for QR rendering. Present for the wechatpay-native channel; absent for stripe and alipay-page.
        providerOrderId:
          type: string
          description: Channel-side order/session identifier created by the payment provider.
        providerId:
          type: string
          enum: [stripe, wechatpay-native, alipay-page]
          description: Payment channel that handled this checkout session.
...